toms feedback applied
